// packages/api/src/generated/models/qualityPredictionModelResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * SmartEM Decisions Backend API
 * API for accessing and managing electron microscopy data
 * OpenAPI spec version: 0.1.1rc42.dev9+g3a5577dc2.d20260427
 */

export interface QualityPredictionModelResponse {
  name: string;
  description: string;
  can_train: boolean;
  can_infer: boolean;
  can_update: boolean;
}